+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This release closes the [0.2.0 milestone](https://github.com/plangrid/ReactiveLi
 ### New
 
 - `TableSectionViewModel` and `CollectionSectionViewModel` now implement `Collection` ([#135](https://github.com/plangrid/ReactiveLists/pull/135), [@benasher44](https://github.com/benasher44))
+- [DifferenceKit](https://github.com/ra1028/DifferenceKit) is now used instead of Dwifft for faster diffing ([#136](https://github.com/plangrid/ReactiveLists/pull/136), [@benasher44](https://github.com/benasher44))
 
 0.1.4
 -----